moved LoginManager sources into LoginDataSource
